requirements.txt
# requirements.txt
Django==5.2.*
djangorestframework==3.14.*
gunicorn>=20.1,<21.0
sphinx>=4.5,<5.0
sphinx-rtd-theme>=1.0,<2.0
psycopg2-binary>=2.9,<3.0
django-simple-history>=3.3,<4.0
dj-database-url>=0.5.0,<1.0
django-cors-headers>=4.7.0,<5.0
django-allauth>=0.57,<66.0
dj-rest-auth==7.0.1 